Adjust delta values
These will be adjusted more sooner or later. Reason is that running
upgrade tests do pass, but some values get smaller deltas when ran
in different maps. This is not really an issue for detection system
overall as values are only few units lower and can be adjusted with some
manual testing. What requires more time is to adjust deltas when
upgrades share a lot of characters, for example 'faster shooting' and
'even faster shooting'.