Fix a bug with beast handler merging
When a beast handler, or actually any tower, is selected the upgrade
panel opens on the opposite side relative to middle coordinate: this was
already well-known fact. But when another tower is selected while first
one has panel open, the panel re-opens without animation and stays on
same side, even if for the latter it would normally appear on the
opposite site. This of course messes upgrading if done immediately after
merging as ocr finds nothing.

Solution is of course very simple: just press esc after merging to close
current planel. Then pressing the target opens the panel on the normal
fashion and won't cause issues.