Modify strong_ocr_check to also return ocr string
Instead of returning just boolean, the detected ocr string is also
included; return value is thus a tuple of type (bool, str). This change
is the used for bot to detect if it's lagging behind in rounds and can
then automatically execute all rounds until it catches up with actual
game round.