Improve adjusting system
Created a _ocr_upgradedata.json to serve as a template for adjusting.
If current upgrade file is not found or has invalid data format, this
template is then copied into temp file. If however current upgrade file
is functional, temp uses this as base.

This change was made in order to make current system more safe, as
previously current data could be lost when error occured during
adjusting. As current data was also the base source for this data, this
would wipe any backup data. Thus, _ocr_upgradedata now fills the role of
this backup whereas .temp_ocr_upgrade_data.json interracts with
upgrades_current.json like in previous system.