First commit
Includes readme.md and all required images. These are used to display
all the readme contents under help window. README is automatically
copied from root folder, if possible. Otherwise uses the existing file.
Images must be copied manually from docs/github/images, but this should
be rare as they are unlikely to change.

To display the markdown in help window, it must be converted to html.
The tkinterweb library used for this annoyingly requires the use of
absolute paths. Thus, each time a help window created, these absolute
paths are written in html file. To avoid storing this data for no
reason, README.html is wiped every time you either
    - close the help window instance, or
    - reopen the main window i.e. start the program
Of course, README.html must also be created again everytime a help
window is reopened, but this is quite a cheap process for such a small
file size.