Add upgrade panel verification
In some rare cases, bot seems to click too fast when another object is
in the way and does not open the upgrade panel. It would then get stuck
until defeat or ocr time limit is reached. This change simply adds the
same check which is already used in placing monkeys: upgrade now checks
the 'sell' text in opened panel. If it does not find it, it attemps a
few more times and if still unsuccesful, clicks on monkey location again
to open the panel if it failed to open.