Improved upgrade check
Bot now performs the 'sell' text check with upgrades similar to when
initially placing them. This is to prevent errors if click does not go
through and bot just searches empty text. Time limit of checking is 10
seconds which should be plenty.

- also made small time adjustment with placement check: in rare cases,
ocr failed to read placed monkey 'sell' message in time and keep trying
to place a monkey when it already exists.
- other minor changes