Update text font, add macOS support
Because setting hotkeys uses pynput.Listener, it had to reworked for
macOS. On this operating system, user must manually write the key to a
popup window. For modifier keys (which are listed in hotkeys info), the
'Key.' must be added in front e.g. Key.ctrl.

Already existing system is still used for non-mac os.